ChatBot/requirements.txt
# Core dependencies
streamlit>=1.32.0
transformers>=4.37.0
torch>=2.2.0
numpy>=1.24.0
chromadb>=0.4.0
langchain-community
langchain

# Database
sqlite3>=3.35.0
chromaDB

# API and networking
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
typing-extensions>=4.9.0

# Model
FastText cc.hi.300.bin

